documentHighlights: stricter symbol bounds
`lsp--document-highlight' now requires the cursor to be within the
symbol, not after it (TES|T() vs TEST|()), for it to keep the symbol
highlights due to the symbol at `point' not changing.

`ccls' highlights the entire macro call when the cursor is within parens
of a macro call, i.e. _TEST|(...)_, but only the macro itself (_TEST_)
if the cursor is on it. As such, moving from TEST()| to TEST|() causes
_TEST()_ to still be highlighted, unlike when moving from | TEST() to
TES|T(). This is inconsistent and a minor inconvenience.