Add `lsp-install-server-for-buffer'
`lsp-install-server' prompts the user to install a server applicable to
the current buffer.

Refactor the code used to prompt the user for a server, used in `lsp'
and `lsp-install-server', into a separate function, `lsp--read-server'.

Add a new function to install a server by filtering `lsp-clients',
`lsp-install-server-1'. Use it in `lsp-install-server-for-buffer' and
`lsp-install-server', refactoring the latter.